rag-api/requirements.txt
fastapi
uvicorn
pydantic
langchain
langchain-google-genai
google-ai-generativelanguage
sentence-transformers
faiss-cpu
pypdf
unstructured
python-dotenv
langchain-community

#uvicorn main:app --reload --port 8000
#python -m uvicorn main:app --reload --port 8000
#python -m uvicorn challenge_agent:app --reload --port 8000